selerio/requirements.txt
Flask==2.3.3
unicorn==2.1.2
Flask-Cors==5.0.0
google-auth==2.38.0
google-genai==1.3.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
websockets==14.2
Werkzeug==3.1.3